pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vfm"
version = "0.1.0"
description = "Vision Foundation Model for Multi-Domain Object Detection"
authors = [{name = "Michael Rushin", email = "michael.r.rushin@gmail.com"}]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pytorch-lightning>=2.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "numpy>=1.21.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "tqdm>=4.64.0",
    "rich>=13.0.0",
    "einops>=0.6.0",
    "timm>=0.9.0",
    "opencv-python>=4.8.0",
    "pillow>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
    "jupyter>=1.0.0",
    "ipywidgets>=8.0.0",
]

[project.scripts]
vfm-train = "vfm.scripts.train:main"
vfm-evaluate = "vfm.scripts.evaluate:main"
vfm-demo = "vfm.scripts.demo:main"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src/vfm --cov-report=term-missing"
